Fix editor being readonly after unmount and remount
When the editor is unmounted (e.g. you close the active editor from
open files section or delete the currently opened file), the editor
was not responsive to changes upon re-mounting. Turns out it was because
of the incomplete editor service implementation. This PR fixes it by
using the existing editor service from Monaco.

fbshipit-source-id: 33934d01