Redesign starter template
This simplifies the starter template, replaces the expo logo with the snack logo, and replaces react-native-elements with react-native-paper.

The current starter template is ok, but I think the content in the card looks out of place compared to the paragraph.
Also the expo logo is old and the white logo isn't properly visible in the editor. It makes more sense to use the snack logo.

I'm using react-native-paper because I work on it and want to promote it :D
It also has TypeScript definitions now, so we'll get autocomplete. I was waiting for it before adding it here.

fbshipit-source-id: 5ecf352