Force preview to reload when SDK version changes
Fixes #2724

Currently we use the URL as the key, so the iframe doesn't re-mount on SDK version change in some scenarios such as snacks saved to the account.
This adds the sdk version to the key to force a re-mount.

I moved the render method to bottom, that's why the diff looks big. The only change is:

```diff
-            key={url}
+            key={url + this.props.sdkVersion}
```

fbshipit-source-id: 78e0bc3