Tune resource requests and limits for Snack
Reduce Snack (web)'s RAM request from 256MiB to 128MiB since it has been using less than that for a month. Also added a limit of 512MiB and 0.8 cores.

Reduced Snackager's CPU request from 1 core to 750 millicores, and its CPU limit from 2 cores to 1500 millicors. Sometimes it seems to use 2 cores, but when it does, it eats up CPU for hours which looks like a bug or possibly abuse. This caps that a little more.

fbshipit-source-id: a218c47